add domain tests

+++ src/main/java/com/dmitrenko/itemmatchingservice/dto/view/NgramView.java
@@ -14,5 +14,5 @@ import java.util.List;
 @ToString
 @NoArgsConstructor
 public class NgramView {
-	private List<Integer> grams;
+	private List<Long> grams;
 }
